+++ .settings/org.springframework.ide.eclipse.beans.core.prefs
